requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
backports.functools-lru-cache==2.0.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
datasets==3.1.0
dill==0.3.8
filelock==3.16.1
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
huggingface-hub==0.30.2
idna==3.10
importlib-metadata==8.5.0
indic-transliteration==2.3.69
joblib==1.4.2
keras==2.13.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.4.0
packaging==24.2
pandas==2.0.3
propcache==0.2.0
protobuf==4.25.6
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
roman==4.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
shellingham==1.5.4
six==1.17.0
tensorboard
tensorboard-data-server
# tensorflow==2.13.1
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.67.1
transformers
typer==0.15.2
typing-extensions==4.5.0
tzdata==2025.2
urllib3==2.2.3
werkzeug==3.0.6
wrapt==1.17.2
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
